requirements.txt
python-dotenv==1.1.0
azure-identity==1.23.0
azure-ai-projects==1.0.0b11
azure-search-documents==11.5.1
azure-storage-blob==12.25.1
azure-ai-formrecognizer
openai==1.82.1
python-docx==1.2.0
PyPDF2==3.0.1
streamlit==1.44.1